refactor: ValidatedStone을 만드는 역할을 Board로 이동
