+++ src/main/kotlin/domain/Board.kt
@@ -1,20 +1,29 @@
 package domain
 
 class Board(private val players: Players, private val stones: Stones) {
-    fun repeatTurn(coordinateReader: CoordinateReader, winEvent: (winner: Player) -> Unit) {
+    fun repeatTurn(coordinateReader: CoordinateReader): Player {
         while (true) {
             val currentTurnPlayer = players.currentTurn()
-            val stone = stones.makeValidatedStone(currentTurnPlayer, coordinateReader)
+            val stone = makeValidatedStone(currentTurnPlayer, coordinateReader)
             stones.place(stone)
-            if (stones.isWinPlace(stone)) {
-                winEvent(currentTurnPlayer)
-                break
-            }
+            if (isWinPlace(stone)) return currentTurnPlayer
         }
     }
+
+    private fun makeValidatedStone(player: Player, coordinateReader: CoordinateReader): Stone {
+        val coordinate = coordinateReader.read(player.color)
+        val stone = Stone(player.color, coordinate)
+        if (!player.validateOmokRule(stones, stone) || !stones.validateDuplicatedCoordinate(stone)) {
+            return makeValidatedStone(player, coordinateReader)
+        }
+        return stone
+    }
+
+    private fun isWinPlace(stone: Stone): Boolean = RenjuRule.findScore(stones, stone) >= WINNING_CONDITION
 
     companion object {
         val BOARD_SIZE = Point(15, 15)
         val BOARD_START_POINT = Point(0, 0)
+        private const val WINNING_CONDITION = 4
     }
 }
